backend/requirements.txt
# FastAPI本体とASGIサーバー
fastapi==0.115.0
uvicorn[standard]==0.30.6

# 環境変数管理
pydantic-settings==2.4.0
python-dotenv==1.0.1

# LangChain + Gemini連携
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0

# Firebase Admin SDK
firebase-admin==6.5.0

# Google Cloud Speech / Text-to-Speech
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.17.0

# マルチパートフォームデータ処理
python-multipart==0.0.9

# HTTPクライアント（非同期）
httpx==0.27.2
